% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Kfairnesstest}
\alias{Kfairnesstest}
\title{Runs tests for 1-n series with repetition. This test is designed to judge the fairness of ranks.}
\usage{
Kfairnesstest(rank, max, method = "Bonferroni", alpha = 0.05)
}
\arguments{
\item{rank}{1-n series with repetition.}

\item{max}{number of n.}

\item{method}{This test will generate many independent p-values according to each pair, Methods are used to adjust the level of significance.}

\item{alpha}{The defaulted value is 0.05.}
}
\value{
The p-value of each one-sided test. Plus, A matrix contains each pair and whether we should reject H0 considering their p-values and the level of significance.
}
\description{
Runs tests for 1-n series with repetition. This test is designed to judge the fairness of ranks.
}
\examples{
library(KuangTest)
test<-c(1,3,3,3,1,1,3,1,3,3,1,3,3,1,1,1,3,3,3,1,3,2,2,2,2,2,2,4,4,2,2,4,4,2,2,4,4,2,2,4,2,4,4,2,2,4,4,2,2,4,4,2,2,4,4,2,2,4,4,2,2,4,4,2)
Kfairnesstest(rank =test, max = 4,method = "BH")
Kfairnesstest(rank =test, max = 4,method = "B")
}
\references{
Friedman J H, Rafsky L C. Multivariate generalizations of the Wald-Wolfowitz and Smirnov two-sample tests[J]. The Annals of Statistics, 1979: 697-717.

Bonferroni, C. E. "Il calcolo delle assicurazioni su gruppi di teste." In Studi in Onore del Professore Salvatore Ortu Carboni. Rome: Italy, pp. 13-60, 1935.

Benjamini, Yoav, and Yosef Hochberg. "Controlling the false discovery rate: a practical and powerful approach to multiple testing." Journal of the Royal statistical society: series B (Methodological) 57.1 (1995): 289-300.
}
